Split rand and mt_rand into separate files
